@@ -6,23 +6,23 @@
   "main": "src/index.tsx",
   "dependencies": {
     "@reduxjs/toolkit": "^1.5.1",
-    "@types/react-redux": "^7.1.9",
-    "@types/react-router": "5.1.7",
-    "@types/react-router-dom": "5.1.5",
-    "firebase": "^8.3.1",
-    "framer-motion": "1.10.3",
+    "firebase": "^9.19.1",
+    "framer-motion": "10.12.2",
     "normalize.css": "8.0.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-redux": "^7.2.1",
-    "react-router": "5.1.2",
-    "react-router-dom": "5.1.2",
-    "react-scripts": "3.4.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router": "6.10.0",
+    "react-router-dom": "6.10.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
-    "@types/react": "16.9.19",
-    "@types/react-dom": "16.9.5",
-    "typescript": "3.7.5"
+    "@types/react": "18.0.37",
+    "@types/react-dom": "18.0.11",
+    "@types/react-redux": "^7.1.9",
+    "@types/react-router": "5.1.20",
+    "@types/react-router-dom": "5.3.3",
+    "typescript": "5.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
