@@ -39,6 +39,9 @@ export const rootReducer = combineReducers<GlobalState>({
   settings: settingsReducer
 });
 
-export const store = createStore(rootReducer);
+export const store = createStore(
+  rootReducer,
+  (window as any).__REDUX_DEVTOOLS_EXTENSION__ && (window as any).__REDUX_DEVTOOLS_EXTENSION__()
+);
 
 export default store;
